1_foundations/requirements.txt
requests
python-dotenv
gradio
pypdf
openai
openai-agents
# ShopAdvisor dependencies
fastapi
uvicorn[standard]
pandas
faiss-cpu
langchain
langchain-community
langchain-openai
sentence-transformers